Extract ConvTranspose op into a separate module
The `ops::conv` module was getting large, so split out ConvTranspose into its
own module.
